[🔧수정] AI28Modal 컴포넌트 리팩토링
