+++ src/containers/AI28Modal/AI28Modal.module.css
@@ -1,5 +1,3 @@
-@import '../../root.css';
-
 .shadow {
   width: 100vw;
   height: 100vh;
@@ -17,6 +15,7 @@
   padding: 24px 48px;
   position: relative;
   border-radius: 20px;
+  background-color: #EAEEFF;
 }
 
 .button {
